Add initial migrations and update models in terapia application
In this commit, key elements of the Terapia project's application are initialized. This includes creating initial migrations for Contacts, Session and Profile models, updating settings for the application, as well as editing the requirements.txt file to ensure project compatibility. Also, the models are registered in the Django admin site.