Refactor code and update models, views, and serializers
The commit includes significant modifications in different areas of the code base. In the models, the maximum length of the "tipo" field was reduced from 100 to 50. The views were heavily refactored to improve the readability and efficiency of the code, and a ProfileViewDetail class was introduced for more specific functionality. The serializers received changes to accommodate the updated models, and allow null values for the "sessao" field. The urls have been modified to match these changes.