Add PacienteSerializer to core serializer
A new class PacienteSerializer was introduced to the core serializer module. This was needed to serialize the patient data attached to profiles. It also includes the get_pacientes method to fetch data about patients linked to a profile. Finally, the field "pacientes" was added to the list of fields in ProfileSerializer's Meta.