Refactor tests and update serializer in 'terapia' app
Test files were refactored, with obsolete files being deleted and a new one created for better organization. The ProfileSerializer in 'terapia' app was updated to include handling for 'pacientes' and 'sessao' data. Changes were also made to view classes including renaming and updating the post methods to accommodate the changes in serializers. This update improves code structure and adds functionality for patient and session data handling.